requirements.txt
django==5.2.12
djangorestframework==3.16.1
psycopg2-binary==2.9.11
pillow==12.1.1
requests==2.32.5
